LayoutTests/fast/js/kde/object_prototype_tostring-expected.txt

PASS _array.toString() is "[object Array]"
PASS _string.toString() is "[object String]"
PASS _boolean.toString() is "[object Boolean]"
PASS _number.toString() is "[object Number]"
PASS _object.toString() is "[object Object]"
PASS _date.toString() is "[object Date]"
PASS _regexp.toString() is "[object RegExp]"
PASS _error.toString() is "[object Error]"
PASS _function.toString() is "[object Function]"
PASS successfullyParsed is true
